requirements.txt
asn1crypto==0.24.0
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
cryptography==2.6.1
devscripts==2.19.5+deb10u1
entrypoints==0.3
Flask==1.0.2
gpg==1.12.0
gunicorn==19.9.0
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.0
pycrypto==2.6.1
pycurl==7.43.0.2
PyGObject==3.30.4
pyinotify==0.9.6
pymongo==3.7.1
pyOpenSSL==19.0.0
PySimpleSOAP==1.16.2
python-apt==1.8.4.2
python-debian==0.1.35
python-debianbts==2.8.2
python-magic==0.4.16
pyxdg==0.25
PyYAML==3.13
reportbug===7.5.3-deb10u1
requests==2.21.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
unidiff==0.5.4
urllib3==1.24.1
Werkzeug==0.14.1
